Update 2017-08-24-day-217.markdown
In #2, either "Kislyak *was* considered . . ." or "Kislyak *has been* considered . . ." The guy has gone home and is probably no longer recruiting spies in DC.

In #3, I think of *Steele* as the private investigator behind the dossier. It could refer to Simpson too of course, but IMO it's ambiguous and confusing at first glance.

In #4, it could be either "set guidelines" or "gave guidance," but "set guidance" isn't idiomatic.